% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dukla}
\alias{dukla}
\title{Lyrics of Dukla}
\format{A data frame with 5 rows and 3 variables:
\describe{
  \item{album}{album}
  \item{song}{song}
  \item{lyrics}{lyrics of the song}
}}
\source{
\url{https://dukla.bandcamp.com/}
}
\usage{
dukla
}
\description{
Lyrics of Czech band Dukla scraped from their bandcamp account.
}
\keyword{datasets}
