% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esazlesa}
\alias{esazlesa}
\title{Lyrics of Esazlesa}
\format{A data frame with 13 rows and 3 variables:
\describe{
  \item{album}{album}
  \item{song}{song}
  \item{lyrics}{lyrics of the song}
}}
\source{
\url{https://esazlesa.bandcamp.com/}
}
\usage{
esazlesa
}
\description{
Lyrics of Czech band Esazlesa scraped from their bandcamp account.
}
\keyword{datasets}
