% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gattaca}
\alias{gattaca}
\title{Lyrics of Gattaca}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{album}{album}
  \item{song}{song}
  \item{lyrics}{lyrics of the song}
}}
\source{
\url{https://gattaca.bandcamp.com/}
}
\usage{
gattaca
}
\description{
Lyrics of Czech band Gattaca scraped from their bandcamp account.
}
\keyword{datasets}
