% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_morphodita.R
\name{gen_morphodita}
\alias{gen_morphodita}
\title{Perform morphological generation}
\usage{
gen_morphodita(data, convert_tagset = "strip_lemma_id", ...)
}
\arguments{
\item{data}{A word for which its forms should be generated (must be lemma)}

\item{convert_tagset}{Apply specified tag set converter
(pdt_to_conll2009 / strip_lemma_comment / strip_lemma_id)}

\item{...}{Other parameters accepted by API}
}
\description{
Perform morphological generation
}
\seealso{
http://lindat.mff.cuni.cz/services/morphodita/api-reference.php
}
