% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plusminusnula}
\alias{plusminusnula}
\title{Lyrics of ±0}
\format{A data frame with 8 rows and 3 variables:
\describe{
  \item{album}{album}
  \item{song}{song}
  \item{lyrics}{lyrics of the song}
}}
\source{
\url{https://plusminusnula.bandcamp.com/}
}
\usage{
plusminusnula
}
\description{
Lyrics of Czech band ±0 scraped from their bandcamp account.
}
\keyword{datasets}
