% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_morphodita.R
\name{recode_tags}
\alias{recode_tags}
\title{Turn the columns from tags into factor and assign them meaningful labels}
\usage{
recode_tags(data)
}
\arguments{
\item{data}{data.frame in which the lemmas with tags are stored}
}
\description{
Turn the columns from tags into factor and assign them meaningful labels
}
\seealso{
https://ufal.mff.cuni.cz/pdt/Morphology_and_Tagging/Doc/hmptagqr.html
}
