% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remek}
\alias{remek}
\title{Lyrics of Remek}
\format{A data frame with 24 rows and 3 variables:
\describe{
  \item{album}{album}
  \item{song}{song}
  \item{lyrics}{lyrics of the song}
}}
\source{
\url{https://remekdiy.bandcamp.com/}
}
\usage{
remek
}
\description{
Lyrics of Czech band Remek scraped from their bandcamp account.
}
\keyword{datasets}
