% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{role}
\alias{role}
\title{Lyrics of Role}
\format{A data frame with 10 rows and 3 variables:
\describe{
  \item{album}{album}
  \item{song}{song}
  \item{lyrics}{lyrics of the song}
}}
\source{
\url{https://rolerolerole.bandcamp.com/}
}
\usage{
role
}
\description{
Lyrics of Czech band Role scraped from their bandcamp account.
}
\keyword{datasets}
