% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_morphodita.R
\name{tag_morphodita}
\alias{tag_morphodita}
\title{Perform morphological analysis of supplied text}
\usage{
tag_morphodita(data, tagset = "pdt_to_conll2009", source = "docker", ...)
}
\arguments{
\item{data}{Text to be analysed}

\item{tagset}{Apply specified tag set converter
(pdt_to_conll2009 / strip_lemma_comment / strip_lemma_id)}

\item{source}{use "docker" or "lindat" for API}

\item{...}{Other parameters accepted by the API (see the API reference)}
}
\description{
Function that send to and get requests from MorphoDiTa API call.
}
\seealso{
http://lindat.mff.cuni.cz/services/morphodita/api-reference.php
}
